Rename env_file ".env" to "qualitas.env" and add .env (#73)
The ".env" file has special meaning in docker compose, it is meant for
defining the defaults for variables used in the docker compose files.
See https://docs.docker.com/compose/environment-variables/#the-env-file

In docker-compose.yml, defining `FLASK_PORT` in `environment` or
`env_file` doesn't work because `environment` and `env_file` only set
environment variables inside the container.  This results in errors
like:

```
$ docker-compose up
WARNING: The FLASK_PORT variable is not set. Defaulting to a blank string.
ERROR: The Compose file './docker-compose.yml' is invalid because:
services.web.ports contains an invalid type, it should be a number, or an object
```

For more information, see https://github.com/docker/compose/issues/4189