Use ajax for getting pypi versions on cnx dashboard
One of the things that causes the cnx dashboard to be very slow is the
requests to pypi to get the versions of the packages.  Change from doing
it in python to using ajax in the browser, that way the page can be
loaded with most of the content ready and the pypi versions are filled
in as the responses come in.